text-summarizer-app/requirements.txt
openai
streamlit
python-dotenv
streamlit